Desktop_app/requirements.txt

PyQt5_sip == 12.8.1

firebase_admin == 4.4.0

openpyxl == 3.0.5

pandas == 1.1.3

pywin32 == 228
